common: add live-build-hooks
This allows image building code to run certain hooks from live-build,
which should make our life easier by leveraging fixes from them [1].

Comes together is a recipe which download a deb and extract it to a
location in the fakemachine. This is required because the environment
fakemachine creates is a minimal busybox environment without apt.

[1] given that someone cares to upgrade the path to the package, which
currently has hard-coded version.